agent-sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ioi-agent"
version = "0.1.0"
description = "The Agent Swarm Framework for the Internet of Intelligence"
readme = "README.md"
requires-python = ">=3.10"
license = "BBSL"
keywords = ["ai", "agent", "blockchain", "ioi", "web4"]
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "pydantic>=2.0.0",
    "cryptography>=41.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ioi_sdk"]
